refactor: auth 관련 파일 구조 이동 및 리팩토링
# Conflicts:
#	payment/src/main/java/com/homeaid/settlement/controller/SettlementController.java
