+++ user/src/main/java/com/homeaid/auth/dto/request/CustomerSignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.homeaid.dto.request;
+package com.homeaid.auth.dto.request;
 
 import com.homeaid.domain.Customer;
 import com.homeaid.domain.enumerate.GenderType;
