+++ user/src/main/java/com/homeaid/auth/security/filter/JwtAuthenticationFilter.java
@@ -1,11 +1,11 @@
-package com.homeaid.security.filter;
+package com.homeaid.auth.security.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.homeaid.auth.service.RefreshTokenService;
+import com.homeaid.auth.security.jwt.JwtTokenProvider;
+import com.homeaid.auth.user.CustomUserDetails;
+import com.homeaid.auth.util.CookieUtil;
 import com.homeaid.dto.request.SignInRequestDto;
-import com.homeaid.security.jwt.JwtTokenProvider;
-import com.homeaid.security.user.CustomUserDetails;
-import com.homeaid.security.util.CookieUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
