+++ user/src/main/java/com/homeaid/auth/service/AuthServiceImpl.java
@@ -1,18 +1,15 @@
-package com.homeaid.security.auth.service;
+package com.homeaid.auth.service;
 
 import com.homeaid.auth.dto.TokenResponse;
 import com.homeaid.auth.exception.TokenErrorCode;
-import com.homeaid.auth.service.RefreshTokenService;
-import com.homeaid.auth.service.TokenBlacklistService;
 import com.homeaid.domain.Customer;
 import com.homeaid.domain.Manager;
 import com.homeaid.dto.request.SignInRequestDto;
 import com.homeaid.exception.CustomException;
 import com.homeaid.exception.UserErrorCode;
 import com.homeaid.repository.UserRepository;
-import com.homeaid.security.jwt.JwtTokenProvider;
-import com.homeaid.security.user.CustomUserDetails;
-import com.homeaid.security.user.CustomUserDetailsService;
+import com.homeaid.auth.security.jwt.JwtTokenProvider;
+import com.homeaid.auth.user.CustomUserDetails;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
