+++ user/src/main/java/com/homeaid/controller/SwaggerAuthController.java
@@ -3,12 +3,12 @@ package com.homeaid.controller;
 import com.homeaid.common.response.CommonApiResponse;
 import com.homeaid.domain.Customer;
 import com.homeaid.domain.Manager;
-import com.homeaid.dto.request.CustomerSignUpRequestDto;
-import com.homeaid.dto.request.ManagerSignUpRequestDto;
+import com.homeaid.auth.dto.request.CustomerSignUpRequestDto;
+import com.homeaid.auth.dto.request.ManagerSignUpRequestDto;
+import com.homeaid.auth.dto.response.SignUpResponseDto;
+import com.homeaid.auth.dto.response.SignInResponseDto;
+import com.homeaid.auth.service.AuthService;
 import com.homeaid.dto.request.SignInRequestDto;
-import com.homeaid.dto.response.SignUpResponseDto;
-import com.homeaid.dto.response.SignInResponseDto;
-import com.homeaid.security.auth.service.AuthService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
